pyproject.toml
[tool.poetry]
name = "pyspark-template"
version = "0.2.0"
description = "a PySpark template project code organization"
authors = ["John Smith <john@example.com>"]
readme = "README.md"
packages = [{include = "pyspark_template"}]

[tool.poetry.dependencies]
python = "^3.10"
pyspark = "^3.5.2"
quinn = "^0.9.0"
pandas = "^1.5.3"
pyarrow = "^10.0.1"
click = "^8.1.3"
kafka-python = "^2.0.2"
retrying = "^1.3.4"



[tool.poetry.group.dev.dependencies]
black = {extras = ["jupyter"], version = "^22.12.0"}
findspark = "^2.0.1"
pytest = "^7.2.1"
pytest-black = "^0.3.12"
pytest-isort = "^3.1.0"
chispa = "^0.9.2"
ipykernel = "^6.20.2"
flake8 = "^6.0.0"
mypy = "^0.991"
docker = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.poetry.scripts]
main = "pyspark_template.jobs.main:main"


